requirements.txt
click==7.1.2
Flask==1.1.2
freeze==3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
six==1.15.0
Werkzeug==1.0.1
Pillow==8.0
numpy>=1.19.2
requests===2.25.0
scikit-image>=0.17.2
imutils>=0.5.3
opencv-python>=4.4.0.46
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scipy==1.5.4
tifffile==2020.11.18
pymongo>=3.11.1
dnspython>=2.0.0
